Initial commit
Breaking out the Curator API from the main Curator module is a
crucial step to making it easier to maintain, and more accessible
to others who want to contribute.

This is not ready for mass consumption yet, but it is derived from
working code, so it won't be long.  The next major release of Curator
will make use of this module.